Move insertAttribute into astHelper.ts
